+++ source/Examples/Resources/AzDoTaggingPermission/1-AddAzDoTaggingPermission.ps1
@@ -1,0 +1,25 @@
+<#
+    .DESCRIPTION
+        This example shows how to deny work item tag creation for a group in Azure DevOps.
+#>
+
+New-AzDoAuthenticationProvider -OrganizationName 'test-organization' -PersonalAccessToken 'my-pat'
+
+Configuration Example
+{
+    Import-DscResource -ModuleName 'AzureDevOpsDscNative'
+
+    node localhost
+    {
+        AzDoTaggingPermission 'AddAzDoTaggingPermission'
+        {
+            Ensure      = 'Present'
+            ProjectName = 'MyProject'
+            GroupName   = '[MyProject]\Contributors'
+            isInherited = $false
+            Permissions = @(
+                @{ Permission = 'Create tag definition'; Access = 'Deny' }
+            )
+        }
+    }
+}