+++ source/Modules/AzureDevOpsDsc.Common/LocalizedData/002.LocalizedDataAzSerializationPatten.ps1
@@ -40,6 +40,15 @@ data LocalizedDataAzSerializationPatten
         # Project Permissions
         # Example: $PROJECT:vstfs:///Classification/TeamProject/78a5065f-3043-426f-9cc5-785748b18f9d
         ProjectPermission = '^\$PROJECT:vstfs:\/{{3}}Classification\/TeamProject\/{0}$'
+        # Tagging Permissions
+        # Example: /78a5065f-3043-426f-9cc5-785748b18f9d
+        TaggingPermission = '^\/{0}$'
+        # Analytics Permissions
+        # Example: $/78a5065f-3043-426f-9cc5-785748b18f9d
+        AnalyticsPermission = '^\$\/{0}$'
+        # AnalyticsViews Permissions
+        # Example: $/Shared/78a5065f-3043-426f-9cc5-785748b18f9d
+        AnalyticsViewsPermission = '^\$\/Shared\/{0}$'
         # Build/Pipeline Permissions  (ProjectId or ProjectId/PipelineId)
         # Example: 78a5065f-3043-426f-9cc5-785748b18f9d  or  78a5065f.../123
         BuildPermission = '^{0}(\/[0-9]+)?$'